obe_assess/assessment_creation/requirements.txt
Django>=4.2
djangorestframework
PyMuPDF          # pip install pymupdf
reportlab
requests
